requirements.txt
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
build==0.7.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
lxml==4.7.1
mysql-connector-python==8.0.27
packaging==21.3
pep517==0.12.0
protobuf==3.19.1
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
tkcalendar==1.6.1
tomli==2.0.0
urllib3==1.26.8
